Make template friendlier for code reloading
Before the change to make the history object a defonce, I was getting this error
in the console:

```
Failed to execute 'write' on 'Document': It isn't possible to write into a
document from an asynchronously-loaded external script unless it is explicitly
opened.
```

I also fixed the on-jsload function to reload updated versions of the var.